requirements.txt
fastapi
uvicorn
python-multipart
open3d
trimesh
torch
numpy
scipy
supabase
python-dotenv
